+++ modules/apps/layout/layout-uad/src/main/java/com/liferay/layout/uad/anonymizer/BaseLayoutSetPrototypeUADAnonymizer.java
@@ -50,11 +50,13 @@ public abstract class BaseLayoutSetPrototypeUADAnonymizer
 	}
 
 	@Override
-	public void delete(LayoutSetPrototype layoutSetPrototype)
+	public void delete(LayoutSetPrototype layoutSetPrototype, long userId)
 		throws PortalException {
 
-		layoutSetPrototypeLocalService.deleteLayoutSetPrototype(
-			layoutSetPrototype);
+		if (layoutSetPrototype.getUserId() == userId) {
+			layoutSetPrototypeLocalService.deleteLayoutSetPrototype(
+				layoutSetPrototype);
+		}
 	}
 
 	@Override
