+++ modules/apps/portal-workflow/portal-workflow-kaleo-uad/src/main/java/com/liferay/portal/workflow/kaleo/uad/anonymizer/BaseKaleoLogUADAnonymizer.java
@@ -48,8 +48,10 @@ public abstract class BaseKaleoLogUADAnonymizer
 	}
 
 	@Override
-	public void delete(KaleoLog kaleoLog) throws PortalException {
-		kaleoLogLocalService.deleteKaleoLog(kaleoLog);
+	public void delete(KaleoLog kaleoLog, long userId) throws PortalException {
+		if (kaleoLog.getUserId() == userId) {
+			kaleoLogLocalService.deleteKaleoLog(kaleoLog);
+		}
 	}
 
 	@Override
