+++ modules/apps/portal-workflow/portal-workflow-kaleo-uad/src/main/java/com/liferay/portal/workflow/kaleo/uad/anonymizer/BaseKaleoTaskInstanceTokenUADAnonymizer.java
@@ -55,11 +55,14 @@ public abstract class BaseKaleoTaskInstanceTokenUADAnonymizer
 	}
 
 	@Override
-	public void delete(KaleoTaskInstanceToken kaleoTaskInstanceToken)
+	public void delete(
+			KaleoTaskInstanceToken kaleoTaskInstanceToken, long userId)
 		throws PortalException {
 
-		kaleoTaskInstanceTokenLocalService.deleteKaleoTaskInstanceToken(
-			kaleoTaskInstanceToken);
+		if (kaleoTaskInstanceToken.getUserId() == userId) {
+			kaleoTaskInstanceTokenLocalService.deleteKaleoTaskInstanceToken(
+				kaleoTaskInstanceToken);
+		}
 	}
 
 	@Override
