+++ modules/apps/portal/portal-uad/src/main/java/com/liferay/portal/uad/anonymizer/BasePortletItemUADAnonymizer.java
@@ -48,8 +48,12 @@ public abstract class BasePortletItemUADAnonymizer
 	}
 
 	@Override
-	public void delete(PortletItem portletItem) throws PortalException {
-		portletItemLocalService.deletePortletItem(portletItem);
+	public void delete(PortletItem portletItem, long userId)
+		throws PortalException {
+
+		if (portletItem.getUserId() == userId) {
+			portletItemLocalService.deletePortletItem(portletItem);
+		}
 	}
 
 	@Override
